python3Packages.joinmarket: allow Python 3.12
Python 3.12 is the default in Nixos 24.11.
All tests succeed with Python 3.12.

Context:

Related issue:
https://github.com/JoinMarket-Org/joinmarket-clientserver/issues/1589

This issue contains no hints on what changes caused joinmarket to be
incompatible/compatible with Python 3.12.

The restriction to Python <3.12 was added
here (https://github.com/JoinMarket-Org/joinmarket-clientserver/pull/1587)
in Oct 2023.

This post
(https://github.com/JoinMarket-Org/joinmarket-clientserver/issues/1589#issuecomment-2119278070)
in the above issue mentions on 2024-05-19 that tests are running again
with Python 3.12.

The current joinmarket version (0.9.11) was released on 2024-02-22.